Move AwsManagedImage out of image.ts into separate files
